Show the products based on filter
Consider a situation where we want to filter the product list based on
the brand and category variables. For example, you want to list all the
products that fall under the category laptop and tablets and from
manufacturers google and dell. With the help of matrix variables, we can
from a URL to bind the brand and category variables' values into the URL